requirements.txt
aiobotocore==2.7.0
aiohttp==3.9.3
aioitertools==0.7.1
aiosignal==1.2.0
alabaster==0.7.12
altair==5.0.1
anyio==4.2.0
appdirs==1.4.4
archspec==0.2.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.8.1
astroid==2.14.2
astropy==5.3.4
asttokens==2.0.5
async-lru==2.0.4
atomicwrites==1.4.0
attrs==23.1.0
autobahn==23.6.2
Automat==20.2.0
autopep8==2.0.4
Babel==2.11.0
bcrypt==3.2.0
beautifulsoup4==4.12.2
binaryornot==0.4.4
black==23.11.0
bleach==4.1.0
blinker==1.6.2
bokeh==3.3.4
boltons==23.0.0
botocore==1.31.64
Bottleneck==1.3.7
Brotli==1.0.9
cachetools==4.2.2
certifi==2024.2.2
cffi==1.16.0
channels==4.1.0
chardet==4.0.0
charset-normalizer==2.0.4
click==8.1.7
cloudpickle==2.2.1
colorama==0.4.6
colorcet==3.1.0
comm==0.2.1
constantly==23.10.4
contourpy==1.2.0
cookiecutter==2.6.0
cryptography==42.0.5
cssselect==1.2.0
cycler==0.11.0
cytoolz==0.12.2
daphne==4.1.2
dask==2023.11.0
datashader==0.16.0
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
deptry==0.16.1
diff-match-patch==20200713
dill==0.3.7
distributed==2023.11.0
distro==1.8.0
Django==5.0.4
django-cors-headers==4.3.1
django-json-widget==2.0.1
django-storages==1.14.2
djangorestframework==3.15.1
docstring-to-markdown==0.11
docutils==0.18.1
entrypoints==0.4
et-xmlfile==1.1.0
executing==0.8.3
fastjsonschema==2.16.2
filelock==3.13.1
flake8==7.0.0
Flask==2.2.5
fonttools==4.25.0
frozenlist==1.4.0
fsspec==2023.10.0
gensim==4.3.2
gitdb==4.0.7
GitPython==3.1.37
greenlet==3.0.1
h5py==3.9.0
HeapDict==1.0.1
holoviews==1.18.3
hvplot==0.9.2
hyperlink==21.0.0
idna==3.4
imagecodecs==2023.1.23
imageio==2.33.1
imagesize==1.4.1
imbalanced-learn==0.11.0
importlib-metadata==7.0.1
incremental==22.10.0
inflection==0.5.1
iniconfig==1.1.1
intake==0.6.8
intervaltree==3.1.0
ipykernel==6.28.0
ipython==8.20.0
ipython-genutils==0.2.0
ipywidgets==7.8.1
isort==5.9.3
itemadapter==0.3.0
itemloaders==1.1.0
itsdangerous==2.0.1
jaraco.classes==3.2.1
jedi==0.18.1
jellyfish==1.0.1
Jinja2==3.1.3
jmespath==1.0.1
joblib==1.2.0
json5==0.9.6
jsonpatch==1.33
jsonpointer==2.1
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
jupyter==1.0.0
jupyter_client==8.6.0
jupyter-console==6.6.3
jupyter_core==5.5.0
jupyter-events==0.8.0
jupyter-lsp==2.2.0
jupyter_server==2.10.0
jupyter_server_terminals==0.4.4
jupyterlab==4.0.11
jupyterlab-pygments==0.1.2
jupyterlab_server==2.25.1
jupyterlab-widgets==1.0.0
keyring==24.3.1
kiwisolver==1.4.4
lazy_loader==0.3
lazy-object-proxy==1.6.0
lckr_jupyterlab_variableinspector==3.1.0
linkify-it-py==2.0.0
llvmlite==0.42.0
lmdb==1.4.1
locket==1.0.0
lxml==4.9.3
lz4==4.3.2
Markdown==3.4.1
markdown-it-py==2.2.0
MarkupSafe==2.1.3
matplotlib==3.8.0
matplotlib-inline==0.1.6
mccabe==0.7.0
mdit-py-plugins==0.3.0
mdurl==0.1.0
mistune==2.0.4
more-itertools==10.1.0
mpmath==1.3.0
msgpack==1.0.3
multidict==6.0.4
multipledispatch==0.6.0
munkres==1.1.4
mypy==1.8.0
mypy-extensions==1.0.0
mysql-connector-python==8.0.33
nbclient==0.8.0
nbconvert==7.10.0
nbformat==5.9.2
nest-asyncio==1.6.0
networkx==3.1
nltk==3.8.1
notebook==7.0.8
notebook_shim==0.2.3
numba==0.59.0
numexpr==2.8.7
numpy==1.26.4
numpydoc==1.5.0
openpyxl==3.0.10
overrides==7.4.0
packaging==23.2
pandas==2.1.4
pandocfilters==1.5.0
panel==1.3.8
param==2.1.0
paramiko==2.8.1
parsel==1.8.1
parso==0.8.3
partd==1.4.1
pathspec==0.10.3
patsy==0.5.3
pexpect==4.8.0
pickleshare==0.7.5
pillow==10.2.0
pip==23.3.1
platformdirs==3.10.0
plotly==5.9.0
pluggy==1.0.0
ply==3.11
prometheus-client==0.14.1
prompt-toolkit==3.0.43
Protego==0.1.16
protobuf==3.17.3
psutil==5.9.0
psycopg2==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyarrow==14.0.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.11.1
pycosat==0.6.6
pycparser==2.21
pyct==0.5.0
pycurl==7.45.2
pydeck==0.8.0
PyDispatcher==2.0.5
pydocstyle==6.3.0
pyerfa==2.0.0
pyflakes==3.2.0
Pygments==2.15.1
pylint==2.16.2
pylint-venv==3.0.3
pyls-spyder==0.4.0
PyNaCl==1.5.0
pyodbc==5.0.1
pyOpenSSL==24.0.0
pyparsing==3.0.9
PyQt5==5.15.10
PyQt5-sip==12.13.0
PyQtWebEngine==5.15.6
PySocks==1.7.1
pytest==7.4.0
python-dateutil==2.8.2
python-json-logger==2.0.7
python-lsp-black==2.0.0
python-lsp-jsonrpc==1.1.2
python-lsp-server==1.10.0
python-slugify==5.0.2
pytoolconfig==1.2.6
pytz==2023.3.post1
pyviz_comms==3.0.2
pywavelets==1.5.0
pywin32-ctypes==0.2.2
PyYAML==6.0.1
pyzmq==25.1.2
QDarkStyle==3.2.3
qstylizer==0.2.2
QtAwesome==1.2.2
qtconsole==5.5.1
QtPy==2.4.1
queuelib==1.6.2
referencing==0.30.2
regex==2023.10.3
requests==2.31.0
requests-file==1.5.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.3.5
rope==1.12.0
rpds-py==0.10.6
Rtree==1.0.1
ruamel.yaml==0.17.21
s3fs==2023.10.0
scikit-image==0.22.0
scikit-learn==1.3.0
scipy==1.11.3
Scrapy==2.11.1
seaborn==0.12.2
Send2Trash==1.8.2
service-identity==18.1.0
setuptools==68.2.2
sip==6.7.12
six==1.16.0
smart-open==5.2.1
smmap==4.0.0
sniffio==1.3.0
snowballstemmer==2.2.0
sortedcontainers==2.4.0
soupsieve==2.5
Sphinx==5.0.2
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
spyder==5.5.1
spyder-kernels==2.5.0
SQLAlchemy==2.0.25
sqlparse==0.5.0
stack-data==0.2.0
statsmodels==0.14.0
sympy==1.12
tables==3.9.2
tabulate==0.9.0
tblib==1.7.0
tenacity==8.2.2
terminado==0.17.1
text-unidecode==1.3
textdistance==4.2.1
threadpoolctl==2.2.0
three-merge==0.1.1
tifffile==2023.4.12
tinycss2==1.2.1
tldextract==3.2.0
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.1
toolz==0.12.0
tornado==6.3.3
tqdm==4.65.0
traitlets==5.7.1
truststore==0.8.0
Twisted==23.10.0
txaio==23.1.1
typing_extensions==4.9.0
tzdata==2023.3
uc-micro-py==1.0.1
ujson==5.4.0
Unidecode==1.2.0
urllib3==2.0.3
w3lib==2.1.2
watchdog==2.1.6
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==0.58.0
Werkzeug==2.2.3
whatthepatch==1.0.2
wheel==0.41.2
whitenoise==6.6.0
widgetsnbextension==3.6.6
win-inet-pton==1.1.0
wrapt==1.14.1
xarray==2023.6.0
xlwings==0.29.1
xyzservices==2022.9.0
yapf==0.40.2
yarl==1.9.3
zict==3.0.0
zipp==3.17.0
zope.interface==5.4.0
zstandard==0.19.0
